+++ src/main/java/com/MMOManagement/SpringApp/Repository/Game/Characters/AtrybutRepository.java
@@ -3,5 +3,5 @@ package com.MMOManagement.SpringApp.Repository.Game.Characters;
 import com.MMOManagement.SpringApp.Model.Game.Characters.Atrybut;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AtrybutRepository  extends JpaRepository<Atrybut, Integer> {
+public interface AtrybutRepository  extends JpaRepository<Atrybut, Long> {
 }
