+++ src/main/java/com/MMOManagement/SpringApp/Repository/Game/Characters/AtrybutyPostaciRepository.java
@@ -3,5 +3,5 @@ package com.MMOManagement.SpringApp.Repository.Game.Characters;
 import com.MMOManagement.SpringApp.Model.Game.Characters.AtrybutyPostaci;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AtrybutyPostaciRepository extends JpaRepository<AtrybutyPostaci, Integer> {
+public interface AtrybutyPostaciRepository extends JpaRepository<AtrybutyPostaci, Long> {
 }
