+++ src/main/java/com/MMOManagement/SpringApp/Repository/Game/Characters/PostacDPSRepository.java
@@ -3,5 +3,5 @@ package com.MMOManagement.SpringApp.Repository.Game.Characters;
 import com.MMOManagement.SpringApp.Model.Game.Characters.PostacDPS;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PostacDPSRepository extends JpaRepository<PostacDPS, Integer> {
+public interface PostacDPSRepository extends JpaRepository<PostacDPS, Long> {
 }
