+++ src/main/java/com/MMOManagement/SpringApp/Repository/Game/Items/PrzedmiotRepository.java
@@ -3,5 +3,5 @@ package com.MMOManagement.SpringApp.Repository.Game.Items;
 import com.MMOManagement.SpringApp.Model.Game.Items.Przedmiot;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PrzedmiotRepository  extends JpaRepository<Przedmiot, Integer> {
+public interface PrzedmiotRepository  extends JpaRepository<Przedmiot, Long> {
 }
