+++ src/main/java/com/MMOManagement/SpringApp/Repository/Users/Moderation/ModeratorCzatuRepository.java
@@ -3,5 +3,5 @@ package com.MMOManagement.SpringApp.Repository.Users.Moderation;
 import com.MMOManagement.SpringApp.Model.Users.Moderation.ModeratorCzatu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ModeratorCzatuRepository extends JpaRepository<ModeratorCzatu,Integer> {
+public interface ModeratorCzatuRepository extends JpaRepository<ModeratorCzatu,Long> {
 }
