+++ src/main/java/com/MMOManagement/SpringApp/Repository/Users/OsobaRepository.java
@@ -3,5 +3,5 @@ package com.MMOManagement.SpringApp.Repository.Users;
 import com.MMOManagement.SpringApp.Model.Users.Osoba;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OsobaRepository  extends JpaRepository<Osoba, Integer> {
+public interface OsobaRepository  extends JpaRepository<Osoba, Long> {
 }
