+++ src/main/java/com/MMOManagement/SpringApp/Repository/Users/Playerbase/PrzywodcaKlanuRepository.java
@@ -3,5 +3,5 @@ package com.MMOManagement.SpringApp.Repository.Users.Playerbase;
 import com.MMOManagement.SpringApp.Model.Users.Playerbase.PrzywodcaKlanu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PrzywodcaKlanuRepository  extends JpaRepository<PrzywodcaKlanu,Integer> {
+public interface PrzywodcaKlanuRepository  extends JpaRepository<PrzywodcaKlanu,Long> {
 }
